Add advertisement warnings
CoreBluetooth specification dictates that only 28 bytes of data are
available for advertising over BLE. This limits the amout of data used
for the local application name.

These warnings are provided to indicate to the user/developer that the
BLE application may not behave as expected if given long names and UUIDs
for advertising.