Adjust BlueZAdvertizements to only advertize the primary service
DBus fails to advertize when registering an advertizement with more than
one 32-bit service UUID. This may be a result of limited space availble
on the advertizement (in addition to the manufacturing information and
the local name data).

Limiting the number of advertized service UUIDs allows for successful
registration when multiple services are used.