Expose MTU onto the peripheral server
This is really a brute force and odd approach to the solution. Each
backend obtains the MTU differently, and because data transfers are
often initiated by the central, the MTU will default to None until some
action is taken. On windows and MacOS, that action must be
a subscription to a characteristic so that the MTU can be negotiated and
exposed. On Linux, it is updated every read and write request.